Move the remaining non-compiler VM code into C++.

+++ vm/AllocTracker.h
@@ -48,7 +48,7 @@ void dvmDisableAllocTracker(void);
         if (gDvm.allocRecords != NULL)                                      \
             dvmDoTrackAllocation(_clazz, _size);                            \
     }
-void dvmDoTrackAllocation(ClassObject* clazz, int size);
+void dvmDoTrackAllocation(ClassObject* clazz, size_t size);
 
 /*
  * Generate a DDM packet with all of the tracked allocation data.
